requirements.txt
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0
pytest==6.2.5
python-dotenv==0.19.0
pyyaml==6.0.1
requests==2.31.0 